requirements.txt
python-telegram-bot==20.5
httpx==0.24.1
python-dotenv==1.0.0